services/trainingservice/requirements.txt
grpcio-tools ~= 1.35
numpy